Move api.js from /style to /src and changed the paths impor paths in Form.jsx and ProtectedRoute.jsx.
