+++ frontend/src/api.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { ACCESS_TOKEN } from "../constants";
+import { ACCESS_TOKEN } from "./constants";
 
 
 const api = axios.create({
