+++ frontend/src/components/Form.jsx
@@ -9,7 +9,7 @@
  */
 
 import { useState } from "react";
-import api from "../styles/api";
+import api from "../api";
 import { useNavigate } from "react-router-dom";
 import { ACCESS_TOKEN, REFRESH_TOKEN } from "../constants";
 import "../styles/Forms.css"
