+++ frontend/src/components/ProtectedRoute.jsx
@@ -10,7 +10,7 @@
 
 import { Navigate } from "react-router-dom";
 import { jwtDecode } from "jwt-decode";
-import api from "../styles/api";
+import api from "../api";
 import { REFRESH_TOKEN, ACCESS_TOKEN } from "../constants";
 import { useState, useEffect } from "react";
